+++ TDEngine2/source/editor/CDevConsoleWindow.cpp
@@ -1,6 +1,7 @@
 #include "../../include/editor/CDevConsoleWindow.h"
 #include "../../include/core/IImGUIContext.h"
 #include "../../include/core/IInputContext.h"
+#include "../../include/core/CGameUserSettings.h"
 #include "../../include/utils/CFileLogger.h"
 #include "stringUtils.hpp"
 
@@ -9,6 +10,57 @@
 
 namespace TDEngine2
 {
+	static std::tuple<std::string, bool> SetConsoleVariableCommand(const std::vector<std::string>& args)
+	{
+		if (args.size() != 2)
+		{
+			return std::make_tuple(Wrench::StringUtils::Format("Expected 2 arguments <var_name> <var_value> for command set, found {0}", args.size()), true);
+		}
+
+		E_RESULT_CODE result = RC_OK;
+
+		const std::string& valueStr = args.back();
+
+		if (CGameUserSettings::Get()->GetVariableInfo(args.front()) != nullptr) // change value because variable already declared
+		{
+			if (valueStr.find_first_of('.') != std::string::npos) // \note try parse float value
+			{
+				result = CGameUserSettings::Get()->SetFloatVariable(args.front(), static_cast<F32>(atof(valueStr.c_str())));
+			}
+			else if (valueStr.find_first_of('\"') != std::string::npos) // \note string
+			{
+				result = CGameUserSettings::Get()->SetStringVariable(args.front(), valueStr.substr(1, valueStr.length() - 2));
+			}
+			else
+			{
+				result = CGameUserSettings::Get()->SetInt32Variable(args.front(), atoi(valueStr.c_str()));
+			}
+		}
+		else
+		{
+			if (valueStr.find_first_of('.') != std::string::npos) // \note try parse float value
+			{
+				result = CGameUserSettings::Get()->CreateFloatVariable(args.front(), Wrench::StringUtils::GetEmptyStr(), E_CONSOLE_VARIABLE_FLAGS::NONE, static_cast<F32>(atof(valueStr.c_str())));
+			}
+			else if (valueStr.find_first_of('\"') != std::string::npos) // \note string
+			{
+				result = CGameUserSettings::Get()->CreateStringVariable(args.front(), Wrench::StringUtils::GetEmptyStr(), E_CONSOLE_VARIABLE_FLAGS::NONE, valueStr.substr(1, valueStr.length() - 2));
+			}
+			else
+			{
+				result = CGameUserSettings::Get()->CreateInt32Variable(args.front(), Wrench::StringUtils::GetEmptyStr(), E_CONSOLE_VARIABLE_FLAGS::NONE, atoi(valueStr.c_str()));
+			}
+		}
+
+		if (RC_OK != result)
+		{
+			return std::make_tuple("Cannot create a variable", true);
+		}
+
+		return std::make_tuple(Wrench::StringUtils::Format("A new variable \"{0}\" successfully created", args.front()), false);
+	}
+
+
 	CDevConsoleWindow::CDevConsoleWindow() :
 		CBaseEditorWindow()
 	{
@@ -29,7 +81,8 @@ namespace TDEngine2
 		mpInputContext = pInputContext;
 
 		// \note Register built-in commands
-		RegisterCommand("clear", [this](auto&&) { ClearHistory(); return Wrench::StringUtils::GetEmptyStr(); });
+		RegisterCommand("clear", [this](auto&&) { ClearHistory(); return std::make_tuple(Wrench::StringUtils::GetEmptyStr(), false); });
+		RegisterCommand("set", SetConsoleVariableCommand);
 
 		mIsInitialized = true;
 
@@ -74,7 +127,9 @@ namespace TDEngine2
 			return RC_FAIL;
 		}
 
-		_writeToLog(iter->second(args));
+		auto&& outputInfo = iter->second(args);
+		_writeToLog(std::get<std::string>(outputInfo), std::get<bool>(outputInfo));
+
 		return RC_OK;
 	}
 
@@ -125,14 +180,14 @@ namespace TDEngine2
 			auto&& tokens = Wrench::StringUtils::Split(mCurrInputBuffer, " ");
 			if (tokens.empty())
 			{
-				_writeToLog("#[Error] Can't evaluate an empty line", true);
+				_writeToLog("Can't evaluate an empty line", true);
 				return;
 			}
 
-			_writeToLog(Wrench::StringUtils::Format("#{0}", mCurrInputBuffer));
+			_writeToLog(mCurrInputBuffer);
 			if (ExecuteCommand(tokens[0], { tokens.cbegin() + 1, tokens.cend() }) != RC_OK)
 			{
-				_writeToLog(Wrench::StringUtils::Format("#[Error] Unrecognized command: {0}", tokens[0]), true);
+				_writeToLog(Wrench::StringUtils::Format("Unrecognized command: {0}", tokens[0]), true);
 			}
 
 			mCurrInputBuffer.clear();
@@ -151,7 +206,7 @@ namespace TDEngine2
 					for (auto iter = mLog.cbegin() + mCurrPos; iter != mLog.cend(); ++iter)
 					{
 						std::tie(message, isErrorMessage) = *iter;
-						mpImGUIContext->Label(message, isErrorMessage ? errorTextColor : mainTextColor);
+						mpImGUIContext->Label(Wrench::StringUtils::Format("# {0} {1}", isErrorMessage ? "[Error]" : Wrench::StringUtils::GetEmptyStr(), message), isErrorMessage ? errorTextColor : mainTextColor);
 					}
 				}
 				mpImGUIContext->EndChildWindow();