+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Now **CGameUserSettings** supports so called console variables. A console variable is an option (integer, float or string) that could be defined through two main approaches:
 using in-game console or defined in settings.cfg file.
 
-- A new method GetDisplaySize was added into API of **IImGUIContext** interface.
+- A new method GetDisplaySize was added into API of **IImGUIContext** interface. Also ImGUI windows now can define font scale.
 
 ## [0.6.23] 2023-12-30